restructure connect directory
