+++ packages/apolloschurchapp/src/tabs/connect/LikedContentFeed/index.js
@@ -1,0 +1,3 @@
+import LikedContentFeed from './LikedContentFeed';
+
+export default LikedContentFeed;
