+++ packages/apolloschurchapp/src/tabs/connect/LikedFeed/LikedFeed.js
@@ -4,7 +4,7 @@ import { Query } from 'react-apollo';
 import BackgroundView from 'apolloschurchapp/src/ui/BackgroundView';
 import FeedView from 'apolloschurchapp/src/ui/FeedView';
 
-import getLikedContent from './getLikedContent';
+import getLikedContent from '../getLikedContent';
 /** A FeedView wrapped in a query to pull content data. */
 class ContentFeed extends PureComponent {
   /** Function for React Navigation to set information in the header. */
