+++ packages/apolloschurchapp/src/tabs/connect/LikedFeed/index.js
@@ -1,0 +1,3 @@
+import LikedFeed from './LikedFeed';
+
+export default LikedFeed;
