+++ packages/apolloschurchapp/src/tabs/connect/UserAvatarHeader/index.js
@@ -1,0 +1,3 @@
+import UserAvatarHeader from './UserAvatarHeader';
+
+export default UserAvatarHeader;
